Predictive Model of Text Messages using R
I have made an analysis of my text messages. I made a word cloud of the
texts. I made a model which would require you to enter a text message
and would give you the list of contacts who could have sent the text.
This model works by comparing the keywords of the entered text and the
texts of my phone.